Added TestCases

+++ Glass-Service/src/main/java/com/cg/gs/controller/GlassController.java
@@ -2,6 +2,8 @@ package com.cg.gs.controller;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,9 +24,6 @@ import com.cg.gs.service.SequenceGeneratorService;
 
 import lombok.extern.slf4j.Slf4j;
 
-//@Restcontroller means it is a combination of controller and RequestBody
-//@Controller means it is handle the web request
-//@RequestMapping means it is handle the map the web request
 @RestController
 @Slf4j
 @RequestMapping("/api/v1")
@@ -38,33 +37,79 @@ public class GlassController {
 
 	@GetMapping("/allglasses") // users/admin
 	public ResponseEntity<List<Glass>> getAllGlasses() throws GlassNotFoundException {
-		// productserviceimpl.getAllProducts();
-		return new ResponseEntity<>(glassServiceImpl.getAllGlass(), HttpStatus.OK);
+		List<Glass> glass=glassServiceImpl.getAllGlass();
+		log.info("starting  of get mapping");
+	
+		if(glass.size()>0) {
+			log.debug("glasses are {}"
+					+ glass);
+		 return new  ResponseEntity<>(glass,HttpStatus.OK); 
+		}
+		else {
+			log.debug(" no glass found ");
+			 return new  ResponseEntity<>(glass,HttpStatus.NO_CONTENT); 
+		}
 	}
 
 	// admin/users
-	@GetMapping("/getglass/{id}")
-	public ResponseEntity<Glass> getGlassById(@PathVariable Integer id) throws GlassNotFoundException {
-		Glass glass = glassServiceImpl.getGlassById(id);
-		return ResponseEntity.ok().body(glass);
+	@GetMapping("/glasses/{id}")
+	public ResponseEntity<Glass> getGlassById(@Valid @PathVariable Integer id) throws GlassNotFoundException {
+		Glass glass= glassServiceImpl.getGlassById(id);
+		if(glass!=null){
+		  return ResponseEntity.ok().body(glass);
+		}
+		  else {
+			return new   ResponseEntity(glass,HttpStatus.NOT_FOUND);
+		  }
 	}
 
 	@PostMapping("/addglass") // only admin
 	public ResponseEntity<Glass> addGlass(@RequestBody Glass glass) throws NoProperDataException {
-		glass.setGlass_id(sequenceGeneratorService.getSequenceNumberForGlass(Glass.SEQUENCE_NAME));
-		// productserviceimpl.addProduct(product);
-		return new ResponseEntity<>(glassServiceImpl.addGlass(glass), HttpStatus.CREATED);
+		if(glass!=null) 
+		{
+			glass.setGlassId(sequenceGeneratorService.getSequenceNumberForGlass(Glass.SEQUENCE_NAME));
+			glassServiceImpl.addGlass(glass);
+			log.error("added glass");
+			return new ResponseEntity<>(glass,HttpStatus.CREATED);
+			
+		}
+		else
+		{
+			throw new NoProperDataException("Please fill fields");
+			
+		}
 	}
 
 	@PutMapping("/updateglass/{id}") // admin only @PutMapping("/updateproduct/{id}")
-	public ResponseEntity<Glass> updateGlass(@RequestBody Glass glass, @PathVariable Integer id)
+	public String updateGlass(@Valid @RequestBody Glass glass)
 			throws GlassNotFoundException {
-		return ResponseEntity.ok(glassServiceImpl.updateGlass(glass, id));
+		String gla=glassServiceImpl.updateGlass(glass);
+		return gla;
 	}
 
 	@DeleteMapping("/deleteglass/{id}") // only delete
 	public ResponseEntity<String> deleteGlass(@PathVariable Integer id) throws GlassNotFoundException {
-		glassServiceImpl.deleteGlass(id);
-		return ResponseEntity.ok(id + " deleted successfully");
+		int count=1;
+		for(int i=1;i>=count;count++)
+		{
+			if(count==1)
+			{
+			try {
+				glassServiceImpl.deleteGlass(id);
+			} catch (GlassNotFoundException e) {
+				throw new GlassNotFoundException("Glass with the id "+id+" doesn't exist");
+				//log.error(e.getMessage());
+			}
+			}
+			else
+			{
+				
+				log.info("id not found");
+			}
+			
+		}
+		
+		
+			return ResponseEntity.ok(" deleted operation done ");
 	}
 }
